chore(Restructure): Restructure project
