+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-native-in-app-notification",
   "version": "2.2.0",
   "description": "Customisable in-app notification component for React Native",
-  "main": "Notification.js",
+  "main": "src/index.js",
   "repository": "https://github.com/robcalcroft/react-native-in-app-notification.git",
   "author": "Rob Calcroft <rob@calcroft.com>",
   "license": "MIT",
@@ -11,8 +11,8 @@
   },
   "peerDependencies": {
     "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "react-native": "^0.43.4"
+    "react": ">=16.3.0",
+    "react-native": ">=0.54.0"
   },
   "dependencies": {
     "react-native-iphone-x-helper": "^1.2.0",
