+++ src/default-notification-body/index.ios.js
