+++ src/notification.js
@@ -3,13 +3,13 @@ import PropTypes from 'prop-types';
 import { Animated, StyleSheet, Image } from 'react-native';
 import { getStatusBarHeight, isIphoneX } from 'react-native-iphone-x-helper';
 
-import DefaultNotificationBody from './DefaultNotificationBody';
+import DefaultNotificationBody from './default-notification-body';
 
 const styles = StyleSheet.create({
   notification: {
     position: 'absolute',
-    width: '100%',
-  },
+    width: '100%'
+  }
 });
 
 class Notification extends Component {
@@ -24,11 +24,19 @@ class Notification extends Component {
 
     this.state = {
       animatedValue: new Animated.Value(0),
-      isOpen: false,
+      isOpen: false
     };
   }
 
-  show({ title, message, onPress, icon, vibrate } = { title: '', message: '', onPress: null, icon: null, vibrate: true }) {
+  show(
+    { title, message, onPress, icon, vibrate } = {
+      title: '',
+      message: '',
+      onPress: null,
+      icon: null,
+      vibrate: true
+    }
+  ) {
     const { closeInterval } = this.props;
     const { isOpen } = this.state;
 
@@ -37,48 +45,51 @@ class Notification extends Component {
     clearTimeout(this.currentNotificationInterval);
 
     const showNotificationWithStateChanges = () => {
-      this.setState({
-        isOpen: true,
-        title,
-        message,
-        onPress,
-        icon,
-        vibrate,
-      }, () => this.showNotification(() => {
-        this.currentNotificationInterval = setTimeout(() => {
-          this.setState({
-            isOpen: false,
-            title: '',
-            message: '',
-            onPress: null,
-            icon: null,
-            vibrate: true,
-          }, this.closeNotification);
-        }, closeInterval);
-      }));
+      this.setState(
+        {
+          isOpen: true,
+          title,
+          message,
+          onPress,
+          icon,
+          vibrate
+        },
+        () => this.showNotification(() => {
+            this.currentNotificationInterval = setTimeout(() => {
+              this.setState(
+                {
+                  isOpen: false,
+                  title: '',
+                  message: '',
+                  onPress: null,
+                  icon: null,
+                  vibrate: true
+                },
+                this.closeNotification
+              );
+            }, closeInterval);
+          })
+      );
     };
 
     if (isOpen) {
-      this.setState(
-        { isOpen: false },
-        () => this.closeNotification(showNotificationWithStateChanges),
-      );
+      this.setState({ isOpen: false }, () => this.closeNotification(showNotificationWithStateChanges));
     } else {
       showNotificationWithStateChanges();
     }
   }
 
-  showNotification(done = () => {}) {
+  showNotification(done) {
     Animated.timing(this.state.animatedValue, {
       toValue: 1,
-      duration: this.props.openCloseDuration,
+      duration: this.props.openCloseDuration
     }).start(done);
   }
 
-  closeNotification(done = () => {}) {
+  closeNotification(done) {
     Animated.timing(this.state.animatedValue, {
       toValue: 0,
-      duration: this.props.openCloseDuration,
+      duration: this.props.openCloseDuration
     }).start(done);
   }
 
@@ -88,18 +99,10 @@ class Notification extends Component {
       topOffset,
       backgroundColour,
       iconApp,
-      notificationBodyComponent: NotificationBody,
+      notificationBodyComponent: NotificationBody
     } = this.props;
 
-    const {
-      animatedValue,
-      title,
-      message,
-      onPress,
-      isOpen,
-      icon,
-      vibrate,
-    } = this.state;
+    const { animatedValue, title, message, onPress, isOpen, icon, vibrate } = this.state;
 
     const height = baseHeight + this.heightOffset;
 
@@ -109,13 +112,15 @@ class Notification extends Component {
           styles.notification,
           { height, backgroundColor: backgroundColour },
           {
-            transform: [{
-              translateY: animatedValue.interpolate({
-                inputRange: [0, 1],
-                outputRange: [-height + topOffset, 0],
-              }),
-            }],
-          },
+            transform: [
+              {
+                translateY: animatedValue.interpolate({
+                  inputRange: [0, 1],
+                  outputRange: [-height + topOffset, 0]
+                })
+              }
+            ]
+          }
         ]}
       >
         <NotificationBody
@@ -139,11 +144,8 @@ Notification.propTypes = {
   height: PropTypes.number,
   topOffset: PropTypes.number,
   backgroundColour: PropTypes.string,
-  notificationBodyComponent: PropTypes.oneOfType([
-    PropTypes.node,
-    PropTypes.func,
-  ]),
-  iconApp: Image.propTypes.source,
+  notificationBodyComponent: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+  iconApp: Image.propTypes.source
 };
 
 Notification.defaultProps = {
@@ -153,7 +155,7 @@ Notification.defaultProps = {
   topOffset: 0,
   backgroundColour: 'white',
   notificationBodyComponent: DefaultNotificationBody,
-  iconApp: null,
+  iconApp: null
 };
 
 export default Notification;
